Add files via upload
This is what I could do with the help of YouTube and VScode suggestions. I made sure to avoid the use of ChatGPT and I was 90% successful and can explain the code and everything that is in it and what they do. For questions text me on my WhatsApp. if you don't have my WhatsApp, email me at templeson04@gmail.com